requirements.txt
click==7.1.1
environs==7.3.1
Flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
marshmallow==3.5.1
python-dotenv==0.12.0
python-magic-bin==0.4.14
Werkzeug==1.0.0
